fix : Service Student, Teacher 분리
